Update codebase for EF Core and .NET 9.0 compatibility
This commit introduces a new configuration file (`efpt.config.json`) for code generation settings and updates the project file (`OneBase.Db.PowerTools.csproj`) to target .NET 9.0, including a reference to the `Microsoft.EntityFrameworkCore.SqlServer` package.

Several model classes have been auto-generated or modified to align with the updated database schema, enhancing the application's data structure. The `OneBaseContext` class has been updated to include new DbSet properties and methods for executing stored procedures asynchronously, improving database interaction functionality.

Overall, these changes aim to enhance maintainability, performance, and modern development practices within the application.

+++ Src/sln/TypeCatch.Net5/DbExt/DbInitializer.cs
@@ -28,7 +28,7 @@
         db.Users.Add(new User { UserId = "Plr3", FullName = "", Note = "Prepopulated", CreatedAt = now });
         db.Users.Add(new User { UserId = "Plr4", FullName = "", Note = "Prepopulated", CreatedAt = now });
 
-        db.AppStngs.Add(new AppStng { LesnTyp = LessonType.PhrasesRandm, SubLesnId = "0", CreatedAt = DateTime.Now, Id = 1, UserId = "Plr1", Note = "Auto pre-loaded." });
+        db.AppStngs.Add(new AppStng { LesnTyp = (int)LessonType.PhrasesRandm, SubLesnId = "0", CreatedAt = DateTime.Now, Id = 1, UserId = "Plr1", Note = "Auto pre-loaded." });
 
         await db.TrySaveReportAsync();
       }
